requirements.txt
django
django-heroku
gunicorn
python-dotenv
requests
python-dateutil